Rebase and squash proto3_presence branch onto internal_v5_0.
Merging is going to produce a complicated history with merge conflicts that are infeasibly difficult to resolve, specifically in the protoc generated code. This is because proto3_presence is based on internal_v4_0 before it was completed and pushed to github. Not only do we have to solve the merge conflicts for the additional API changes since we branched off, but we also need to resolve the changes for GAPICv2, and the squash merge to github master. Suffice to say that rebasing and squashing is easier to resolve, review and track in the history.

Change-Id: I05e0139843fe641979fdfc5f1756f3c18b441724