Add method to extract all errors of a partial failure status at once
This is useful because getting all errors via their operation index is inefficient. In #getGoogleAdsErrors(long, GoogleAdsFailureT) the error paths for all failures are build just to pick those for the requested operation index. That leads creation of many ErrorPath objects for all not requested operation indexes which. Those are immediately discarded what leads to avoidable GC pressure.

This can already be achieved with the current API, but I think it belongs into the error util. The code (Java 11) I currently use as supplement:
status.getDetailsList().stream()
  .map(errorUtils::getGoogleAdsFailure)
  .map(errorUtils::getGoogleAdsErrors)
  .flatMap(List::stream)
  .collect(toSet());